refactor(legacy-filter-panel): rename to original file name
