+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterPanel.js
